+++ apps/dashboard/src/app/account/settings/getAccount.ts
@@ -1,8 +1,8 @@
 import { API_SERVER_URL } from "@/constants/env";
 import type { Account } from "@3rdweb-sdk/react/hooks/useApi";
 import { getAuthToken } from "../../api/lib/getAuthToken";
-import { isOnboardingComplete } from "../../login/isOnboardingRequired";
 import { loginRedirect } from "../../login/loginRedirect";
+import { isOnboardingComplete } from "../../login/onboarding/isOnboardingRequired";
 
 /**
  * Just get the account object without enforcing onboarding.
