+++ apps/dashboard/src/app/login/onboarding/Title.tsx
@@ -1,4 +1,5 @@
 import type { JSX } from "react";
+
 type TitleAndDescriptionProps = {
   heading: string | JSX.Element;
   description: string | JSX.Element;
