--- apps/dashboard/src/components/onboarding/types.ts
@@ -1,8 +1,0 @@
-import type { Team } from "../../@/api/team";
-
-export type OnboardingScreen =
-  | { id: "onboarding" }
-  | { id: "linking" }
-  | { id: "confirming" }
-  | { id: "confirmLinking" }
-  | { id: "plan"; team: Team };
